restAPI/rapi/requirements.txt
asgiref==3.5.0
certifi==2021.10.8
distlib==0.3.4
Django==4.0.4
django-cors-headers==3.11.0
djangorestframework==3.13.1
djangorestframework-jwt==1.11.0
djangorestframework-simplejwt==5.1.0
filelock==3.6.0
pipenv==2022.4.8
platformdirs==2.5.1
PyJWT==1.7.1
pytz==2022.1
six==1.16.0
sqlparse==0.4.2
tzdata==2022.1
virtualenv==20.14.1
virtualenv-clone==0.5.7
